# openapi/spec.openapi.yml
openapi: 3.0.0
servers:
  - description: OBADA Staging API
    url: https://dev.api.obada.io
  - description: OBADA Production API
    url: http://api.obada.io
info:
  version: "0.0.1"
  title: OBADA API
  termsOfService: 'https://www.obada.io/terms/'
  contact:
    name: Obada Dev Team
    email: techops@obada.io
    url: 'https://www.obada.io'
...
paths:
...
  /api/client/obit:
    post:
      summary: Save Client Obit
      description: ''
      operationId: saveClientObit
      tags:
        - Helper
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocalObit'
      responses:
        "200":
          $ref: '#/components/responses/ClientObitResponse'
...
  /obit/checksum:
    post:
      summary: Generates the obit checksum.
      description: ''
      operationId: checksum
      tags:
        - Obit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Obit'
      responses:
        "200":
          $ref: '#/components/responses/ChecksumResponse'
        "422":
          $ref: "#components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /obit/id:
    post:
      summary: Generate Obit ID
      description: 'Returns the Obit ID for a given device_id, part_number and serial_number input.'
      operationId: generateId
      tags:
        - Obit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestObitId'
      responses:
        "200":
          $ref: "#/components/responses/ObitIdResponse"
        "422":
          $ref: "#components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    RequestObitId:
      $ref: "definitions/Obit.yml#/RequestObitId"
...
    ClientObit:
      $ref: "definitions/Obit.yml#/ClientObit"
    LocalObit:
      $ref: "definitions/Obit.yml#/LocalObit"
    Obit:
      $ref: "definitions/Obit.yml#/Obit"
...
  responses:
    ObitIdResponse:
      description: Obit ID response
      content:
        application/json:
          schema:
            $ref: "definitions/Obit.yml#/ObitId"
...
    ClientObitResponse:
      description: Client obit response
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: integer
              obit:
                $ref: '#/components/schemas/ClientObit'
...
    ChecksumResponse:
      description: Obit checksum response
      content:
        application/json:
          schema:
            $ref: "definitions/Obit.yml#/Checksum"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "Errors.yml#/InternalServerError"
...